Added tracking scaffolding
