+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "aery"
-version = "0.3.1"
+version = "0.4.0"
 edition = "2021"
 authors = ["iiYese iiyese@outlook.com"]
 repository = "https://github.com/iiYese/aery"
@@ -15,7 +15,6 @@ readme = "README.md"
 bevy = "0.11"
 smallvec = "1.11.0"
 aery_macros = { path = "macros", version = "0.2.0-dev" }
-seq-macro = "0.3.5"
 aquamarine = "0.3.2"
 
 [features]
